backend/tests/performance/requirements.txt
locust>=2.15.1
pytest-locust>=2.4.1
pandas>=1.5.0
faker>=15.3.0
python-dotenv>=0.21.0

